hydro_src/requirements.txt
asgiref==3.8.1
Django==3.1.5
django-environ==0.11.2
django-leaflet==0.27.1
djangorestframework==3.12.2
djangorestframework-gis==0.17
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
whitenoise==6.7.0
django-lockdown==4.0.0